Remove unused Test component and refactor HabitDetail component
