+++ src/App.jsx
@@ -18,7 +18,6 @@ import { Toaster } from "react-hot-toast";
 // Auth
 import { useAuth } from "./context/AuthContext";
 import Landing from "./pages/Landing";
-import Test from "./pages/Test";
 
 const toastOptions = {
   style: {
@@ -40,7 +39,6 @@ function App() {
         <Route element={<ProtectedRoutes />}>
           <Route path="/" element={<Habits />} />
           <Route path="/habits/:id" element={<HabitDetail />} />
-          <Route path="/test" element={<Test />} />
         </Route>
         <Route path="*" element={<NotFound />} />
         <Route path="/login" element={<Login />} />
